wip: add verifier FFI bindings
