+++ rust/pact_verifier_cli/Cargo.toml
@@ -25,7 +25,15 @@ lazy_static = "1.4.0"
 rand = "0.7.3"
 regex = "1.3.9"
 tokio = { version = "0.2.9", features = ["rt-core", "macros"] }
+env_logger = "0.7.1"
+futures = "0.3"
 
 [dev-dependencies]
 quickcheck = "0.9.2"
 expectest = "0.12.0"
+
+[lib]
+crate-type = ["cdylib", "staticlib", "rlib"]
+
+[build-dependencies]
+cbindgen = "0.14.1"
